backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
ecdsa==0.19.1
idna==3.10
pyasn1==0.6.1
python-dotenv==1.1.0
python-jose==3.5.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
transbank-sdk>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
reportlab==4.4.1
openpyxl==3.1.5
celery==5.5.3
redis==6.2.0